+++ src/tests/integration/conftest.py
@@ -103,11 +103,18 @@ def load_taxi_data(engine):
     df.to_sql(name=table_name, con=engine, chunksize=100_000, if_exists="replace")
 
 
+def load_numeric_data(engine):
+    table_name = "numbers"
+    df = pd.DataFrame({"numbers_elements": [1, 2, 3]})
+    df.to_sql(name=table_name, con=engine, chunksize=100_000, if_exists="replace")
+
+
 @pytest.fixture(scope="session")
 def setup_postgreSQL():
     engine = create_engine(_get_database_url("postgreSQL"))
     # Load taxi_data
     load_taxi_data(engine)
+    load_numeric_data(engine)
     yield engine
     engine.dispose()
 
@@ -129,6 +136,7 @@ def setup_mySQL():
     engine = create_engine(_get_database_url("mySQL"))
     # Load taxi_data
     load_taxi_data(engine)
+    load_numeric_data(engine)
     yield engine
     engine.dispose()
 
@@ -150,6 +158,7 @@ def setup_mariaDB():
     engine = create_engine(_get_database_url("mariaDB"), pool_recycle=1800)
     # Load taxi_data
     load_taxi_data(engine)
+    load_numeric_data(engine)
     yield engine
     engine.dispose()
 
@@ -171,7 +180,7 @@ def setup_SQLite():
     engine = create_engine(_get_database_url("SQLite"))
     # Load taxi_data
     load_taxi_data(engine)
-
+    load_numeric_data(engine)
     yield engine
     engine.dispose()
 
@@ -193,7 +202,7 @@ def setup_duckDB():
     engine = create_engine(_get_database_url("duckDB"))
     # Load taxi_data
     load_taxi_data(engine)
-
+    load_numeric_data(engine)
     yield engine
     engine.dispose()